refractor: modularizing code for the save and edit screen
